setup-admin-4-phan-quyen

+++ app/Http/Controllers/Admin/AdminController.php
@@ -7,11 +7,13 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\RegAdminRequest;
 use App\Http\Requests\UploadImageRequest;
 use App\Models\Admin;
+use App\Models\Permission;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use Mockery\Exception;
+use Illuminate\Support\Facades\Gate;
 
 class AdminController extends Controller
 {
@@ -22,7 +24,14 @@ class AdminController extends Controller
      */
     public function index()
     {
-        return response()->json(Admin::orderByDesc('id')->paginate(10));
+        if (request()->user()->can('manage-admin.index')) {
+            return response()->json(Admin::orderByDesc('id')->paginate(10));
+        }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 
     /**
@@ -33,29 +42,37 @@ class AdminController extends Controller
      */
     public function store(RegAdminRequest $request)
     {
-        DB::beginTransaction();
-        try {
-            $request->validated();
+        if (request()->user()->can('manage-admin.store')) {
+            DB::beginTransaction();
 
-            $admin = Admin::create(
-                $request->all([
-                    'user_name','full_name','email','phone_number'
-                ]) + [
-                    'password' => Hash::make($request->get('password')),
-                    'avatar' => Helper::saveImage(null, $request->file('avatar'), 'admin/avata')
-            ]);
-            DB::commit();
+            try {
+                $request->validated();
 
-            return response()->json([
-                $admin->fresh()
-            ]);
-        } catch (Exception $exception) {
-            DB::rollBack();
+                $admin = Admin::create(
+                    $request->all([
+                        'user_name','full_name','email','phone_number'
+                    ]) + [
+                        'password' => Hash::make($request->get('password')),
+                        'avatar' => Helper::saveImage(null, $request->file('avatar'), 'admin/avata')
+                ]);
+                DB::commit();
 
-            return response()->json([
-                'error' =>  $exception
-            ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
+                return response()->json([
+                    $admin->fresh()
+                ]);
+            } catch (Exception $exception) {
+                DB::rollBack();
+
+                return response()->json([
+                    'error' =>  $exception
+                ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
+            }
         }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 
     /**
@@ -66,9 +83,21 @@ class AdminController extends Controller
      */
     public function show($id)
     {
-        $admin = Admin::findOrFail($id);
+        if (request()->user()->can('manage-admin.show')) {
+            $admin = Admin::findOrFail($id);
+            $admin = $admin->load([
+                'role' => function ($query) {
+                    $query->with('permissions')->get();
+                }
+            ]);
 
-        return response()->json($admin);
+            return response()->json($admin);
+        }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 
     /**
@@ -80,33 +109,40 @@ class AdminController extends Controller
      */
     public function update(RegAdminRequest $request, $id)
     {
-        DB::beginTransaction();
+        if (request()->user()->can('manage-admin.update')) {
+            DB::beginTransaction();
 
-        try {
-            $data = $request->all();
-            $admin = Admin::findOrFail($id);
+            try {
+                $data = $request->all();
+                $admin = Admin::findOrFail($id);
 
-            if ($data['password'] == null) {
-                $data['password'] = $admin->password;
-            } else {
-                $data['password'] = bcrypt($data['password']);
+                if ($data['password'] == null) {
+                    $data['password'] = $admin->password;
+                } else {
+                    $data['password'] = bcrypt($data['password']);
+                }
+
+                $data['avatar'] = Helper::saveImage($admin->avatar, $request->file('avatar'), 'admin/avata');
+                $admin->update($data);
+                $admin->save();
+                DB::commit();
+
+                return response()->json([
+                    $admin->fresh()
+                ]);
+            } catch (Exception $exception) {
+                DB::rollBack();
+
+                return response()->json([
+                    'error' =>  $exception
+                ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
             }
-
-            $data['avatar'] = Helper::saveImage($admin->avatar, $request->file('avatar'), 'admin/avata');
-            $admin->update($data);
-            $admin->save();
-            DB::commit();
-
-            return response()->json([
-                $admin->fresh()
-            ]);
-        } catch (Exception $exception) {
-            DB::rollBack();
-
-            return response()->json([
-                'error' =>  $exception
-            ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
         }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 
     /**
@@ -117,35 +153,49 @@ class AdminController extends Controller
      */
     public function destroy($id)
     {
-        $admin = Admin::findOrFail($id);
+        if (request()->user()->can('manage-admin.destroy')) {
+            $admin = Admin::findOrFail($id);
 
-        return response()->json([
-            'success' => $admin->delete()
-        ]);
+            return response()->json([
+                'success' => $admin->delete()
+            ]);
+        }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 
     public function updateImage(UploadImageRequest $request, $id)
     {
-        DB::beginTransaction();
+        if (request()->user()->can('manage-admin.upload-image')) {
+            DB::beginTransaction();
 
-        try {
-            $data = $request->all();
-            $admin = Admin::findOrFail($id);
-            $data['avatar'] = Helper::saveImage($admin->avatar, $request->file('avatar'), 'admin/avata');
+            try {
+                $data = $request->all();
+                $admin = Admin::findOrFail($id);
+                $data['avatar'] = Helper::saveImage($admin->avatar, $request->file('avatar'), 'admin/avata');
 
-            $admin->update($data);
-            $admin->save();
-            DB::commit();
+                $admin->update($data);
+                $admin->save();
+                DB::commit();
 
-            return response()->json([
-                $admin->fresh()
-            ]);
-        } catch (Exception $exception) {
-            DB::rollBack();
+                return response()->json([
+                    $admin->fresh()
+                ]);
+            } catch (Exception $exception) {
+                DB::rollBack();
 
-            return response()->json([
-                'error' =>  $exception
-            ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
+                return response()->json([
+                    'error' =>  $exception
+                ], JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
+            }
         }
+
+        return response([
+            'status' => false,
+            'message' => trans('auth.guard_message'),
+        ], 200);
     }
 }
